requirements.txt
asgiref==3.5.0
#backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.47.0
django-autocomplete-light==3.8.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-grappelli==3.0.2
django-model-utils==4.2.0
django-widget-tweaks==1.4.12
gunicorn==20.0.4
idna==3.3
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
whitenoise==5.2.0